Clean up codes for review
