+++ packages/front-end/enterprise/hooks/useAIChat/types.ts
@@ -68,11 +68,45 @@ export interface UseAIChatOptions {
   getConversationEndpoint?: (conversationId: string) => string;
 
   /**
+   * Returns the URL for cancelling an active stream on the server.
+   * When set, the Cancel button sends a POST here before aborting the local
+   * fetch — this lets the backend distinguish explicit cancels from navigation
+   * disconnects and continue generating in the background for the latter.
+   */
+  getCancelEndpoint?: (conversationId: string) => string;
+
+  /**
    * Called once the POST to `endpoint` returns a successful response (before
    * the response body is read). Use to refresh conversation lists after the
    * server has persisted the user message.
    */
   onStreamAccepted?: () => void;
+
+  /**
+   * Called after a conversation is loaded from the server (initial load or
+   * poll). Receives the raw response so consumers can extract extra fields
+   * (e.g. feedback) without changing the hook's core state.
+   */
+  onConversationLoaded?: (data: unknown) => void;
+
+  /**
+   * Called when the AI response stream completes successfully.
+   * `durationMs` is the wall-clock time from send to stream end.
+   * `toolCallCount` is the number of tool calls the agent made during this turn.
+   */
+  onMessageComplete?: (info: {
+    durationMs: number;
+    toolCallCount: number;
+  }) => void;
+
+  /**
+   * Called when the user explicitly cancels generation.
+   * `durationMs` is the wall-clock time from send to cancel.
+   */
+  onMessageCancelled?: (info: { durationMs: number }) => void;
+
+  /** Called when a non-abort error occurs during message send or streaming. */
+  onMessageError?: (info: { errorType: string; httpStatus?: number }) => void;
 }
 
 export interface ConversationSummary {
@@ -83,6 +117,7 @@ export interface ConversationSummary {
   isStreaming: boolean;
   /** Truncated text of the first user message, for sidebar preview. */
   preview: string;
+  model?: string;
 }
 
 /** GET /chat/:id — messages plus whether the agent is still generating. */
@@ -110,6 +145,8 @@ export interface UseAIChatReturn {
   newChat: () => void;
   loadConversation: (id: string) => Promise<void>;
   loading: boolean;
+  /** True only while fetching historical messages for a conversation (not AI generation). */
+  isLoadingConversation: boolean;
   /** True only while this tab is actively reading an SSE stream from sendMessage. */
   isLocalStream: boolean;
   waitingForNextStep: boolean;
